Name room based on member, if only one member present
If only one member is found in a room, set the room name to
`Unnamed ($MEMBERNAME)`.

If more than one member is found the room name is left as null and the
raw Matrix address is used (as previous behaviour).

Lazy loading is used so member list is not necessarily exhaustive.
The only situation I can imagine this failing is when only one other
user has been messaging in a group chat for a long time.

To guarantee all members are retrived a request to /members for a specified room would be required (not implemented).

Closes #6